tatu-hook/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tatu-hook"
version = "0.2.0"
description = "Security hooks for the Tatu DevSecOps platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Laboratório Hacker" },
    { name = "Julio Silveira Melo" },
]
keywords = ["claude-code", "security", "devsecops", "hooks", "secrets-detection", "sast", "lgpd", "gdpr", "soc2", "fedramp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
yara = ["yara-python>=4.5"]
dev = ["pytest>=8.0"]

[project.urls]
Homepage = "https://github.com/laboratorio-hacker/tatu"
Repository = "https://github.com/laboratorio-hacker/tatu"
Issues = "https://github.com/laboratorio-hacker/tatu/issues"
Documentation = "https://github.com/laboratorio-hacker/tatu#tatu-hook-cli"

[project.scripts]
tatu-hook = "tatu_hook.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
